Split ResolvePending into multiple functions
